+++ sdk/keyvault/azure_security_keyvault_secrets/src/lib.rs
@@ -4,7 +4,6 @@
 #![doc = include_str!("../README.md")]
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
 
-#[expect(deprecated, reason = "requires emitter update")]
 mod generated;
 mod resource;